GCAM_JOB

GCAM_JOB

Usage

GCAM_JOB(GW_activity, method = NULL)

Arguments

GW_activity

Output from GCAM_GW()

method

Specify method for construction job. "Net": in a given year, a technology pre-mature retirement and newly installed net out. "Total": in a given year, a technology pre-mature retirement and newly installed can happen at the same time.

Value

A data frame of all power sector direct job by subsector

Examples

JOB_activity <- GCAM_JOB(GW_activity, "Net")
#> Error in OM_JOB(GW_activity): object 'prj' not found
JOB_activity <- GCAM_JOB(GW_activity)
#> Error in OM_JOB(GW_activity): object 'prj' not found