ReadGCAM

ReadGCAM

Usage

ReadGCAM(
  filetype,
  input_path,
  db_name,
  query = system.file("extdata", "my_batch.xml", package = "GCAMUSAJobs"),
  scen_name,
  prj_name
)

Arguments

filetype

specify the input is a database (folder) or a project (.dat)

input_path

specify the full path for the input file

db_name

specify the database name

query

use the default query file in the package "my_batch.xml"

scen_name

specify the scenario name

prj_name

specify the project (.dat) name either to name it or load the existed one

Value

a list to be called in other functions

Examples

prj <- ReadGCAM(filetype = 'db',
input_path = "path",
db_name = "DB",
query = system.file("extdata", "my_batch.xml", package = "GCAMUSAJobs"),,
scen_name = "Scenario_from_DB",
prj_name = "A_Project_Name.dat")
#> Error in ReadGCAM(filetype = "db", input_path = "path", db_name = "DB",     query = system.file("extdata", "my_batch.xml", package = "GCAMUSAJobs"),     , scen_name = "Scenario_from_DB", prj_name = "A_Project_Name.dat"): unused argument (alist())